Changes: 1. added create tables .yaml files 2. data type modifications in counterparty and item

+++ src/main/java/com/erp_demo/dto/ItemDto.java
@@ -5,7 +5,7 @@ import com.erp_demo.entity.ProductEntity;
 public class ItemDto {
 
     private ProductEntity product;
-    private double price;
+    private Float price;
 
     // constructor
     public ItemDto() {}
@@ -19,11 +19,11 @@ public class ItemDto {
         this.product = product;
     }
 
-    public double getPrice() {
+    public Float getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(Float price) {
         this.price = price;
     }
 }
